aimbot-backend/pyproject.toml
[project]
name = "aimbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.6"
dependencies = [
    "aiohttp>=3.13.2",
    "aiolimiter>=1.2.1",
    "beautifulsoup4>=4.14.2",
    "fastapi[standard]>=0.121.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dateutil>=2.9.0.post0",
]
